GraphQL mutation updated. Insert new user implementation improved. API folder removed. Methods folder renamed to Functions
