--- src/api/interceptors.js
@@ -1,16 +1,0 @@
-import axios from 'axios'
-import { getLoggedUser } from '../helpers/local-storage'
-
-axios.interceptors.request.use(request => {
-  const userData = getLoggedUser()
-
-  userData && (request.headers['Authorization'] = `Bearer ${userData.token}`)
-  return request
-})
-
-axios.interceptors.response.use(
-  response => response,
-  error => error.response?.data || { message: 'Something went wrong, please try it again later' }
-)
-
-export default axios
