+++ src/functions/encrypt.js
